% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{tidy_app}
\alias{tidy_app}
\title{A Shiny app to format R code}
\usage{
tidy_app()
}
\description{
Run a Shiny app that formats R code via \code{\link{tidy_source}()}. This app
uses input widgets, such as checkboxes, to pass arguments to
\code{tidy_source()}.
}
\examples{
if (interactive()) formatR::tidy_app()
}
