% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_pipe}
\alias{tidy_pipe}
\title{Substitute the \pkg{magrittr} pipe with R's native pipe operator}
\usage{
tidy_pipe()
}
\description{
Parse the R code in the RStudio editor, identify \code{\%>\%}, and substitute
with \code{|>}.
}
\note{
Currently this function only works inside the RStudio IDE, and may be
  extended in future to deal with arbitrary R code elsewhere.
}
\examples{\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
formatR::tidy_pipe()
\dontshow{\}) # examplesIf}
}
